Resolving early termination due to protocol URL
The protocol-relative URL was causing this script to terminate early on "compilation." As of December 2014, [Paul Irish's blog](https://www.paulirish.com/2010/the-protocol-relative-url/) on protocol-relative URLs says "2014.12.17: Now that SSL is encouraged for everyone and doesn’t have performance concerns, this technique is now an anti-pattern. If the asset you need is available on SSL, then always use the https:// asset."

My observation is that static resources served through Yotpo are already secure and thus should be enforced using https instead of protocol-relative urls, anyways.